Renames and changes after review
