+++ prawns/identities/src/handlers/identity.clear.identity.extrinsic.ts
@@ -1,6 +1,6 @@
 import { ExtrinsicHandlerContext } from '@subsquid/substrate-processor';
 import { decodeAddress } from '../utils';
-import { SubstrateIdentity, SubstrateNetwork, IdentityAction } from '../model';
+import { SubstrateIdentity, SubstrateNetwork, SubstrateIdentityAction } from '../model';
 import { getManager } from 'typeorm';
 
 export default (network: SubstrateNetwork) =>
@@ -23,7 +23,7 @@ export default (network: SubstrateNetwork) =>
       current: true, // the last set_identity call we get is the current one
       blockNumber,
       date,
-      action: IdentityAction.KILL,
+      action: SubstrateIdentityAction.CLEAR,
     });
 
     await ctx.store.save(identityModel);
